added test cases and reverted difference
